+++ crates/bcr-ebill-transport/src/handler/mod.rs
@@ -7,15 +7,15 @@ use mockall::automock;
 
 use super::{EventEnvelope, EventType};
 
+mod bill_action_event_handler;
 mod bill_chain_event_handler;
 mod bill_chain_event_processor;
 mod bill_invite_handler;
-mod public_bill_chain_event_handler;
 
+pub use bill_action_event_handler::BillActionEventHandler;
 pub use bill_chain_event_handler::BillChainEventHandler;
 pub use bill_chain_event_processor::BillChainEventProcessor;
 pub use bill_invite_handler::BillInviteEventHandler;
-pub use public_bill_chain_event_handler::PublicBillChainEventHandler;
 
 #[cfg(test)]
 impl ServiceTraitBounds for MockNotificationHandlerApi {}
